AS-320: Agrega implementación de la consulta por los datos del resumen de tiendas

+++ Source/Locompro/Models/ViewModels/ShoppingListSummaryStoreVm.cs
@@ -10,7 +10,7 @@ public class ShoppingListSummaryStoreVm
 
     public int ProductsAvailable { get; set; }
     
-    public float PercentageProductsAvailable { get; set; }
+    public int PercentageProductsAvailable { get; set; }
     
     public int TotalCost { get; set; }
 }
